+++ test/int/sql/test_explain_buckets.py
@@ -1,0 +1,9 @@
+from framework.sqltester import (
+    ClusterSingleInstance,
+    sql_test_file,
+)
+
+
+@sql_test_file("explain_buckets.sql")
+class TestExplainBuckets(ClusterSingleInstance):
+    pass