+++ test/int/sql/test_explain_options.py
@@ -1,0 +1,10 @@
+from framework.sqltester import (
+    ClusterSingleInstance,
+    sql_test_file,
+)
+
+
+# Tests various combinations of explain facets
+@sql_test_file("explain_options.sql")
+class TestExplainOptions(ClusterSingleInstance):
+    pass